Fix native image build, and move POJO implementation to dedicated package.
