+++ java/src/main/java/com/powsybl/dataframe/impl/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Implementation of dataframe input/output as POJOs.
+ *
+ * @author Sylvain Leclerc <sylvain.leclerc at rte-france.com>
+ */
+package com.powsybl.dataframe.impl;
